+++ src/prf.rs
@@ -1,0 +1,20 @@
+use rand_core::RngCore;
+use serde::{de::DeserializeOwned, Serialize};
+use std::{error::Error, fmt::Debug};
+use threshold_bls::{
+    group::{Element, Scalar},
+    poly::Poly,
+    sig::{Share},
+};
+use crate::poprf::poprf::POPRF;
+use crate::POPRFError;
+use crate::poprf::Scheme;
+
+pub trait PRFScheme: Scheme {
+    type Error: Error;
+
+    /// Evaluates the PRF on the given plaintext tag and message input.
+    ///
+    /// Will result in the same value as calling `blind_msg`, `blind_eval`, `unblind_resp` in sequence.
+    fn eval(private: &Self::Private, tag: &[u8], msg: &[u8]) -> Result<Vec<u8>, Self::Error>;
+}
